placement_cell/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
iniconfig==2.0.0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pyngo==2.0.1
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
